# book-connect-ui/src/app/services/models/item.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { AccessInfo } from '../models/access-info';
import { SaleInfo } from '../models/sale-info';
import { VolumeInfo } from '../models/volume-info';
export interface Item {
  accessInfo?: AccessInfo;
  id?: string;
  saleInfo?: SaleInfo;
  selfLink?: string;
  volumeInfo?: VolumeInfo;
}