# book-connect-ui/src/openApi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenApi specification",
    "description": "OpenApi documentation for Spring security",
    "contact": {
      "name": "Siddharth",
      "email": "siddharthxlodhi@gmail.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Local ENV"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/googleBooks/search": {
      "get": {
        "tags": [
          "GoogleBook"
        ],
        "operationId": "searchBooks",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startIndex",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "maxResults",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "orderBy",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseItem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AccessInfo": {
        "type": "object",
        "properties": {
          "country": {
            "type": "string"
          },
          "viewability": {
            "type": "string"
          },
          "embeddable": {
            "type": "boolean"
          },
          "publicDomain": {
            "type": "boolean"
          },
          "textToSpeechPermission": {
            "type": "string"
          },
          "webReaderLink": {
            "type": "string"
          },
          "accessViewStatus": {
            "type": "string"
          },
          "quoteSharingAllowed": {
            "type": "boolean"
          }
        }
      },
      "ImageLinks": {
        "type": "object",
        "properties": {
          "smallThumbnail": {
            "type": "string"
          },
          "thumbnail": {
            "type": "string"
          }
        }
      },
      "IndustryIdentifier": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "identifier": {
            "type": "string"
          }
        }
      },
      "Item": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "volumeInfo": {
            "$ref": "#/components/schemas/VolumeInfo"
          },
          "saleInfo": {
            "$ref": "#/components/schemas/SaleInfo"
          },
          "accessInfo": {
            "$ref": "#/components/schemas/AccessInfo"
          },
          "selfLink": {
            "type": "string"
          }
        }
      },
      "PageResponseItem": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Item"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
      "Price": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "format": "double"
          },
          "currencyCode": {
            "type": "string"
          }
        }
      },
      "SaleInfo": {
        "type": "object",
        "properties": {
          "country": {
            "type": "string"
          },
          "saleability": {
            "type": "string"
          },
          "listPrice": {
            "$ref": "#/components/schemas/Price"
          },
          "retailPrice": {
            "$ref": "#/components/schemas/Price"
          },
          "buyLink": {
            "type": "string"
          },
          "ebook": {
            "type": "boolean"
          }
        }
      },
      "VolumeInfo": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "subtitle": {
            "type": "string"
          },
          "authors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "publisher": {
            "type": "string"
          },
          "publishedDate": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "industryIdentifiers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IndustryIdentifier"
            }
          },
          "pageCount": {
            "type": "integer",
            "format": "int32"
          },
          "printType": {
            "type": "string"
          },
          "categories": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "averageRating": {
            "type": "number",
            "format": "double"
          },
          "ratingsCount": {
            "type": "integer",
            "format": "int32"
          },
          "maturityRating": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "previewLink": {
            "type": "string"
          },
          "infoLink": {
            "type": "string"
          },
          "canonicalVolumeLink": {
            "type": "string"
          },
          "imageLinks": {
            "$ref": "#/components/schemas/ImageLinks"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT auth ",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}